+++ app/build.gradle
@@ -30,8 +30,9 @@ dependencies {
     implementation "ru.terrakok.cicerone:cicerone:5.0.0"
 
     //DI
-    implementation 'com.google.dagger:dagger:2.13'
-    annotationProcessor 'com.google.dagger:dagger-compiler:2.13'
+    implementation 'com.google.dagger:dagger:2.16'
+    kapt 'com.google.dagger:dagger-compiler:2.16'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.16'
 
     //moxy
     implementation 'tech.schoolhelper:moxy-x-androidx:1.5.6'
