+++ app/src/main/AndroidManifest.xml
@@ -9,18 +9,21 @@
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
+            android:name=".PartyApp"
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
             android:usesCleartextTraffic="true">
-        <activity android:name=".view.LoginActivity">
+
+        <activity android:name=".presentation.feature.auth.activity.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
         <activity
-                android:name=".view.BottomNavActivity"
+                android:name=".presentation.feature.main.activity.MainActivity"
                 android:theme="@style/AppTheme">
         </activity>
     </application>
