+++ app/src/main/java/com/example/partycheckapp/data/party/Purchase.kt
@@ -1,0 +1,10 @@
+package com.example.partycheckapp.data.party
+
+data class Purchase(
+    val id: Long,
+    val title: String,
+    val price: Int,
+    val userList: ArrayList<Long>,
+    val image: ByteArray,
+    val check: ByteArray
+)
