--- app/src/main/java/com/example/partycheckapp/di/module/AppModule.kt
@@ -1,16 +1,0 @@
-package com.example.partycheckapp.di.module
-
-import android.app.Application
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class AppModule(private val app: Application) {
-
-    @Provides
-    @Singleton
-    fun provideContext(): Context = app.baseContext
-
-}
