+++ app/src/main/java/com/example/partycheckapp/domain/party/MathLogic.kt
@@ -1,0 +1,9 @@
+package com.example.partycheckapp.domain.party
+
+import com.example.partycheckapp.data.party.Party
+
+class MathLogic {
+    var party: Party? = null
+
+
+}
