--- app/src/main/java/com/example/partycheckapp/entity/Purchase.kt
@@ -1,10 +1,0 @@
-package com.example.partycheckapp.entity
-
-data class Purchase(
-    val id: Long,
-    val title: String,
-    val price: Int,
-    val userList: ArrayList<Long>,
-    val image: ByteArray,
-    val check: ByteArray
-)
