--- app/src/main/java/com/example/partycheckapp/model/MathLogic.kt
@@ -1,9 +1,0 @@
-package com.example.partycheckapp.model
-
-import com.example.partycheckapp.entity.Party
-
-class MathLogic {
-    var party: Party? = null
-
-
-}
