+++ app/src/main/java/com/example/partycheckapp/presentation/di/module/AppModule.kt
@@ -1,0 +1,16 @@
+package com.example.partycheckapp.presentation.di.module
+
+import android.app.Application
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class AppModule(private val app: Application) {
+
+    @Provides
+    @Singleton
+    fun provideContext(): Context = app.baseContext
+
+}
