+++ app/src/main/java/com/example/partycheckapp/presentation/di/module/PartyModule.kt
@@ -1,0 +1,8 @@
+package com.example.partycheckapp.presentation.di.module
+
+import dagger.Module
+
+@Module
+class PartyModule {
+
+}
