+++ app/src/main/java/com/example/partycheckapp/presentation/di/scope/PartyScope.kt
