--- app/src/main/java/com/example/partycheckapp/view/PartyApp.kt
@@ -1,18 +1,0 @@
-package com.example.partycheckapp.view
-
-import android.app.Application
-import ru.terrakok.cicerone.Cicerone
-import ru.terrakok.cicerone.Router
-import ru.terrakok.cicerone.NavigatorHolder
-
-class PartyApp : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        INSTANCE = this
-    }
-
-    companion object {
-        lateinit var INSTANCE: PartyApp
-    }
-}
