--- app/src/main/res/layout/activity_bottom_nav.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/rl_bottom_nav"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".view.BottomNavActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/appbar"
-            android:layout_width="match_parent"
-            android:background="@color/colorPrimary"
-            android:theme="@style/Base.Theme.AppCompat.Light.DarkActionBar"
-            android:layout_height="wrap_content">
-
-        <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar"
-                app:titleTextColor="@color/colorWhite"
-                android:layout_width="match_parent"
-                android:layout_height="?actionBarSize">
-
-        </androidx.appcompat.widget.Toolbar>
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <FrameLayout
-            android:id="@+id/container"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_above="@+id/navigation"/>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-            android:id="@+id/navigation"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:itemBackground="@color/colorPrimary"
-            app:itemTextColor="@color/colorPrimaryText"
-            app:itemIconTint="@color/colorPrimaryText"
-            android:layout_alignParentBottom="true"
-            app:menu="@menu/bottom_nav"/>
-
-
-</RelativeLayout>
